optimize

+++ src/main/java/com/example/todo/service/impl/TodoServiceImpl.java
@@ -1,11 +1,15 @@
 package com.example.todo.service.impl;
 
+import com.example.todo.dto.CreateTodoRequestDto;
+import com.example.todo.dto.TodoDTO;
 import com.example.todo.entity.Category;
 import com.example.todo.entity.Status;
 import com.example.todo.entity.Todo;
 import com.example.todo.repository.TodoRepository;
+import com.example.todo.security.CurrentUser;
 import com.example.todo.service.TodoService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -47,13 +51,49 @@ public class TodoServiceImpl implements TodoService {
     }
 
     @Override
-    public void remove(int id) {
+    public ResponseEntity<?> remove(int id, CurrentUser currentUser) {
+        Optional<Todo> byId = todoRepository.findById(id);
+        if (byId.isEmpty() || byId.get().getUser().getId() != currentUser.getUser().getId() || !todoRepository.existsById(id))
+            return ResponseEntity.noContent().build();
         todoRepository.deleteById(id);
+        return ResponseEntity.notFound().build();
     }
 
     @Override
     public Todo save(Todo entity) {
         return todoRepository.save(entity);
+    }
+
+    @Override
+    public Todo save(TodoDTO todoDTO, int id) {
+        Optional<Todo> byId = todoRepository.findById(id);
+        if (byId.isEmpty()) {
+            return null;
+        }
+        Todo todo = byId.get();
+        if (todoDTO.getTitle() != null && !todoDTO.getTitle().isEmpty()) {
+            todo.setTitle(todoDTO.getTitle());
+        }
+        if (todoDTO.getStatus() != null) {
+            todo.setStatus(todoDTO.getStatus());
+        }
+        if (todoDTO.getCategory() != null) {
+            todo.setCategory(todoDTO.getCategory());
+        }
+        return todo;
+    }
+
+    @Override
+    public Todo save(CreateTodoRequestDto createTodoRequestDto, CurrentUser currentUser) {
+        if (currentUser != null) {
+            return Todo.builder()
+                    .title(createTodoRequestDto.getTitle())
+                    .category(createTodoRequestDto.getCategory())
+                    .status(Status.NOT_STARTED)
+                    .user(currentUser.getUser())
+                    .build();
+        }
+        return null;
     }
 
     @Override
